Restructure to InlineRegions
