--- src/main/java/uk/ac/cam/cl/dtg/isaac/dos/content/InlineChoice.java
@@ -1,5 +1,0 @@
-package uk.ac.cam.cl.dtg.isaac.dos.content;
-
-public class InlineChoice extends Choice {
-    public InlineChoice() {}
-}
